Refactor, implement config driven, and support all of the objects for a single OCI client
This is the first version that I actually think anyone else would be interested in.

In this version I implemented a single class for all of the Data Safe objects we can
find and delete and parameterized them. To have the class find and delete more objects you
shouldn't have to write any code - instead you just add the definition of the object up top.

Still more to do to make this work for other Clients but this is a good start