Fix for issue #2
Under the old code I didn't initialize a client for the home region unless it was in the specified list of regions. Which was probably some sort of dumb optimization idea I had as initializing such a client incurs minimal cost.

This change moves that code up a few lines and takes it out of the "else" block so the client will be initialized first before the regional ones.