Added threading support
We now run deletes for a particular grouping of object in every region at the same time. So for example all of the compute instances in all regions, then all block storage in all regions, etc.

Which means more threads = more faster, but only up to the number of regions. Beyond that we won't actually spawn any more threads.