Implement Data Flow and Data Science
These implementations are lightly tested but I think they're correct. Of note is that for Data Science we delete all of the child resources but don't wait for them to complete before trying to delete the Project. This may cause a failure deleting the Project so I will likely improve this in the future.