Tweak to genAI agents for region restrictions
genAI Agents service is available in a limited set of regions. So to avoid needing to wait for errors this code restricts cleaning them up only occur in the regions I know they're in.

This is suboptimal since it means I have to remember to tweak this code as the service rolls out to other regions.

TODO: Figure out a better way!